Backend/requirements.txt
Django==4.2.19
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
drf_yasg==1.21.7
 celery==5.3.6
 redis==5.0.7